+++ wgpu-types/src/tokens.rs
@@ -3,7 +3,7 @@ use macro_rules_attribute::derive;
 use crate::{link_to_wgpu_item, ConstDefault};
 
 /// Token of the user agreeing to access experimental features.
-#[derive(Debug, ConstDefault!, Copy, Clone)]
+#[derive(Debug, ConstDefault!, Copy, Clone, Eq, PartialEq)]
 pub struct ExperimentalFeatures {
     enabled: bool,
 }